rest-graph.rb: add RestGraph#broken_old_rest, see:
http://www.nivas.hr/blog/2010/09/03/facebook-php-sdk-access-token-signing-bug/

if you're getting this error from calling old_rest:

  The method you are calling or the FQL table you are querying cannot be
  called using a session secret or by a desktop application.

then try broken_old_rest instead. the problem is that the access_token
should be formatted by "#{app_id}|#{secret}" instead of the usual one,
and the json returned by facebook is broken too, need to remove the
quotes and the extra '\'. for instance:

  "{\"app_id\":\"123\"}"

instead of the correct one:

  {"app_id":"123"}

so, there are two hacks in broken_old_rest. one is faking access_token,
the other one is fix the json by removing extra quotes and the '\'.