rails_util.rb: we might want to delete old cookie in other cases:
but not in rest_graph_check_cookie! it's ok since it is not calling
rest_graph_write_rg_fbs. we do only need to write our fbs when it is
not inside the cookie, and when we need to write, we definitely want
the new fbs instead of the old one.